+++ src/leiningen/bouncer/types/config.clj
@@ -9,6 +9,11 @@
   ".bouncer/config.edn")
 
 
+(def ^:const backup-config-file
+  "The secondary name of the configuration file."
+  ".wallbrew/bouncer/config.edn")
+
+
 (spec/def ::disabled
   (st/spec
     {:type        :boolean
@@ -53,6 +58,15 @@
      :description (str "The license information for the project. "
                        "If the rule is disabled, a reason must be recorded. "
                        "For example, \"Inherited Eclipse License from source of fork\"")}))
+
+
+(spec/def ::namespace-sorting
+  (st/spec
+    {:type        :map
+     :spec        ::rule
+     :description (str "Fixes the sorting of `:require` blocks in namespace declarations. "
+                       "If the rule is disabled, a reason must be recorded. "
+                       "For example, \"Stateful components must be imported in a predetermined order\"")}))
 
 
 (spec/def ::plugin-name
@@ -113,10 +127,19 @@
      :description "The rules against project.clj that bouncer can enforce and their configuration."}))
 
 
+(spec/def ::fixes
+  (st/spec
+    {:type        :map
+     :spec        (spec/keys :opt-un [::comment
+                                      ::namespace-sorting])
+     :description "The rules against project.clj that bouncer can enforce and their configuration."}))
+
+
 (spec/def ::config
   (st/spec
     {:type        :map
      :spec        (spec/keys :opt-un [::comment
+                                      ::fixes
                                       ::project-rules])
      :description "The configuration for Bouncer."}))
 
@@ -124,16 +147,21 @@
 (def default-config
   "The default configuration for Bouncer."
   {:comment       "The default configuration for Wall Brew projects."
+   :fixes         {:namespace-sorting {:disabled false
+                                       :reason   "Namespace sorting is required for Wall Brew projects."}}
    :project-rules {:license {:comment  "https://github.com/Wall-Brew-Co/open-source?tab=readme-ov-file#licensing"
                              :disabled false}
                    :plugins {:disabled          false
                              :available-plugins [{:comment        "https://github.com/Wall-Brew-Co/rebroadcast?tab=readme-ov-file#sealog-configuration"
                                                   :plugin-name    "com.wallbrew/lein-sealog"
                                                   :plugin-version "1.6.0"}
+                                                 {:comment        "https://github.com/Wall-Brew-Co/rebroadcast?tab=readme-ov-file#bouncer-configuration"
+                                                  :plugin-name    "com.wallbrew/bouncer"
+                                                  :plugin-version "1.6.0"}
                                                  {:comment        "https://github.com/Wall-Brew-Co/rebroadcast?tab=readme-ov-file#clj-kondo-configuration"
                                                   :plugin-name    "com.github.clj-kondo/lein-clj-kondo"
                                                   :plugin-version "2024.08.29"}
-                                                 {:plugin-name    "lein-cljsbuild"
+                                                 {:plugin-name    "lein-cljsbuild/lein-cljsbuild"
                                                   :plugin-version "1.1.8"
                                                   :disabled       true
                                                   :reason         "cljsbuild is only required for ClojureScript and cross-platform projects."}